Report CSSProperty exposure correctly with/without ExecutionContext
A CSSProperty which is behind a disabled runtime flag, and that is
also associated with a origin trial will currently claim to have
CSSExposure::kWeb if Exposure() is called. This is potentially
dangerous, since we could expose something by accident.

This problem happens because we have two virtual functions:

 (1) Exposure(), which ignores origin trials.
 (2) Exposure(const ExecutionContext*), which includes origin trials.

Origin-trial-dependent properties override (2), but not (1), and
(1) returns kWeb by default.

This CL refactors such that there is only one Exposure function to
override, thus solving the problem once and for all.

Bug: 1375994
Change-Id: I65166c68cccfa90d11263c9eb12a276afd6f754a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4042188
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074567}
NOKEYCHECK=True
GitOrigin-RevId: 7f7215f2a47e08b2d224c6cc6260b39facade6e1