Fail fast quiet chip.
The feature is enabled by default.

This CL finalizes a permission request callback as soon as it is clear
that the permission request will be shown as a quiet chip. The quiet
chip is not resolved but is shown as normal. Permission request result
is delivered via PermissionStatus listener.

A permission request can be finalized only if a requesting RFH
subscribed to PermissionStatus onchange event. In case of multiple
duplicated permission requests, all corresponding RFHs should listen
to the change event.

Bug: 1364526
Change-Id: If416a24747cae66316c42ae9c2a89b5d15ee5ccc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3907729
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Ravjit Uppal <ravjit@chromium.org>
Commit-Queue: Illia Klimov <elklm@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Javier Fernandez <jfernandez@igalia.com>
Reviewed-by: Kamila Hasanbega <hkamila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074659}
NOKEYCHECK=True
GitOrigin-RevId: e406ecc1fb52f1a4d9775bfa59ef7a0bafb85050