[editing] (2/4) Make ComparePositions() faster
This patch is the second part of a series of patches to make
`ComparePositions()` faster.

This patch does refactoring to introduces `Comparator` class to
encapsulate traversal methods for help functions:
  * `CompareNodesInSameParent()`
    It will be called three places in the fourth patch.
  * `FindChildInAncestors()`
    It is called in two places
and introduces `Result` enum for ease of understanding the logic of
`ComparePositions()` and ease of applying changes to these functions.

This patch introduces runtime flag `FastComparePositions` to disable
this change.

The series is
  1. https://crrev.com/c/4039247 Move to "compare_positions.cc"
  2. https://crrev.com/c/4040525 Move logic to functions
  3. https://crrev.com/c/4040521 CompareNodesInSameParent faster
  4. https://crrev.com/c/4040842 Relative positions faster

Bug: 1323777
Change-Id: I686e1b520ffef2a4700a5da2efb64fd2722abb01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4040525
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Auto-Submit: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074667}
NOKEYCHECK=True
GitOrigin-RevId: 53d08650e5a603075d8cbf8819b66c6c39bd5e43