Implement RuntimeFeatureStateOverrideContext getters/setters
This CL begins to fill out the RuntimeFeatureStateOverrideContext class
by implementing the getters and setters.

It also begins setting up for the eventual inclusion as a member in
ExecutionContext.

While we are generating getters/setters for features which use
enterprise policies, support for the policies is not yet complete.

Bug: 1377000
Change-Id: I41f811c91a41a377030fa06bdfb7ce040c1eb2df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4034933
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074709}
NOKEYCHECK=True
GitOrigin-RevId: 4ae5d96a6eb27307cf305c98b121f7387e0aefd1