[Reland#2] VT: Fix snapshot rect for a resizing virtual keyboard
==Reland note==

Previous CL (https://crrev.com/c/4021408) was reverted due
to flakes. This CL includes a number of fixes:

  * Correctly measuring and waiting for the keyboard accessory
  * Switching from Android system screenshots to Viz Surface copies
  * Hiding scrollbars, which animate non-deterministically on Android.
  * Allow imperceptible color differences in pixel tests due to
   compositing/raster differences.

PS#1 is the previous CL. Cleanups related to the no-longer-used-here
ForceRedraw API have been split out into https://crrev.com/c/4035007.

==============

ViewTransitions compute the "snapshot rect" which is a viewport-like
rect that doesn't change size when UI such as scrollbars, virtual
keyboard, mobile URL bar are showing or hidden.

When the keyboard is in a resize-content mode (so that innerHeight is
changed by the keyboard, see:
https://developer.chrome.com/blog/viewport-resize-behavior/), the
snapshot rect must account for the keyboard by outsetting the layout
viewport.

This CL plumbs the virtual-keyboard height into the renderer via
SynchronizeVisualProperties so that DocumentTransitionStyleTracker can
use this value to perform the outset. This value includes any keyboard
accessories such as the autofill bar.

Since the keyboard resize behavior is implemented in the //chrome layer,
this CL adds end-to-end chrome browser tests for this behavior. This
requires introducing some hand-rolled pixel testing functionality into
java tests which isn't common so also requires plumbing the
InsertVisualStateCallback mechanism into java, allowing tests to wait on
a Blink main frame to be presented to the screen.

Bug: 1350910
Change-Id: Ie929869d5ed70b478549a4fcc504c8223e173777
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4031356
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1074789}
NOKEYCHECK=True
GitOrigin-RevId: de28eb384848668d988985ab98cdb86533d54034