[editing] (3/4) Make ComparePositions() faster
This patch is the third part of a series of patches to make
`ComparePositions()` faster.

This patch changes `CompareNodesInSameParent()` to reduces number of
iterations of child nodes to
  * The minimum is min(A, B, |A-B|, N-A, N-B)
  * The maximum N/4 where N is number of children
    where A and B are offsets of child nodes, N is number of children
from min(A, B).

This makes `ComparePositions()` faster when A and B are near start/end
of children list and |A-B| is small.

This patch introduces the performance test for selection at end of large
container.

The perf_test[1] results on local machine are:
  Before: 989.820ms (before `FastPositionIterator`)
  Before: 861.820ms (after `FastPositionIterator`)
  After:  126.309ms
It is 6.823 times faster than before.

[1] third_party/blink/perf_tests/editing/many_spans_with_move_backward_character.html

The series is
  1. https://crrev.com/c/4039247 Move to "compare_positions.cc"
  2. https://crrev.com/c/4040525 Move logic to functions
  3. https://crrev.com/c/4040521 CompareNodesInSameParent faster
  4. https://crrev.com/c/4040842 Relative positions faster

Bug: 1323777
Change-Id: I9ec035c1d34179622f6f31aea479c3c6c814011e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4040521
Auto-Submit: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075022}
NOKEYCHECK=True
GitOrigin-RevId: bb35f761e9ba5c5b0c1497ee3820027802017209