WebNN: Implement the synchronous build method of MLGraph
This CL implements MLGraph::BuildSync() that is called by MLGraphBuilder
to implement the buildSync() method for the dedicated worker.
BuildSync() has the similar usage and function as MLGraph::BuildAsync().
The difference is if there are no validation errors, it calls
BuildSyncImpl() implemented by an MLGraph backend, such as
MLGraphXnnpack, and builds the platform specific graph synchronously in
the caller's thread.

This CL also introduces the MLGraphBuilder::BackendForTesting and
value-parameterized MLGraphTestBase that help test MLGraph and its
sub-classes for both asynchronous and synchronous execution modes.

Bug: 1273291
Change-Id: I3a98d92c7b8f3817f84c96e19685270031e96aa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4035256
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Commit-Queue: ningxin hu <ningxin.hu@intel.com>
Cr-Commit-Position: refs/heads/main@{#1075051}
NOKEYCHECK=True
GitOrigin-RevId: c1b722504f4151cec0c75f1208c121253c4f8a9f