Add tests for interactions trace events
Events are dispatched with the name "EventTiming" and enable the
Interactions track in DevTools Perf panel.

The test triggers a keyboard interaction and waits for the
PerformanceEventTiming API to report the corresponding entries in order
to collect the matching trace events that will be tested.

Bug: 1383130
Change-Id: Ia431009316663e4e61a840abd193254a8cb1751f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4047407
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075149}
NOKEYCHECK=True
GitOrigin-RevId: 095075e0963b26483f91e6eda3d526354498c10d