Ensure WFWI::current_drag_data_ is null at the end of a drag.
This is a tentative fix for the DCHECK bugs below.  The bugs appeared at
two different times, but they share the common symptom that
WebFrameWidgetImpl::DragTargetDragEnter very infrequently encounters a
non-null current_drag_data_.  We were not able to repro the problem.

This tentative fix assumes that an early return on a drag-leave may
be accidentally skipping setting the state to null.

Bug: 1316301, 1367848
Change-Id: Ib6e9e000271515a0659a6780a59467db149d3b3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645670
Commit-Queue: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075157}
NOKEYCHECK=True
GitOrigin-RevId: 88da1e1c25b97c0e40e5fe9dbe60e00c9d66b0f5