Define class ScriptPromiseResultTracker for metric tracking
As part of the effort to improve metric and latency tracking of Chrome
RTC APIs, we have introduced the class ScriptPromiseResultTracker which wraps around ScriptPromiseResolver.

In the follow-up CLs, we would replace the usage of ScriptPromiseResolver with ScriptPromiseResultTracker for our APIs of interest.

Bug: 1373398
Change-Id: Ifa91ff641da1c703ec0f701a743c867af0203591
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4031635
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Palak Agarwal <agpalak@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075160}
NOKEYCHECK=True
GitOrigin-RevId: 272c037d9e40f966c4bcf4f92c3c02b2414c2d56