Use single-decision point on what type of AXObject to create, if any
This helps clean up redundancies in the code and allows us to use
one call to determine the type of AXObject to create. It also helps
with the downstream CL that will somewhat separate the decision of
which type of object to create from what it is backed by (node or layout object).

Another benefit of this CL is that it fixes cases where direct
children of a <head> that is display locked were somethings being
created and triggering assertions found elsewhere in the code, as found by chrishtr.

NOTRY=true

Bug: none
Change-Id: I879ac0e4b10b79c1997dae0d6afaa6ee58c1bd37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4043566
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075277}
NOKEYCHECK=True
GitOrigin-RevId: e58920864058dd65e44d9811ee0104755e7a5267