Prerender: Add first implementation of prerender-in-new-tab mode
This CL adds first end-to-end implementation of the prerender-in-new-tab
mode behind a new feature flag (kPrerender2InNewTab). The mode is
available via "target_hint" parameter specified on speculation rules
(this part was implemented by the previous CL).

Also, this CL adds browser tests to make sure the behavior in basic
cases. Follow-up CLs will add more tests (including WPTs) to cover edge
cases.

> Implementation Design

See the design document for details:
https://docs.google.com/document/d/1ZSwrsBtwgxO8xtgX0hA23YR92H62Y0o60ByVO352RL4/edit?usp=sharing

This CL can roughly be divided into 2 parts: triggering part and
activation part.

1) Triggering
When "target_hint: _blank" is specified on speculation rules,
PrerenderHostRegistry creates PrerenderNewTabHandle. The handle is
responsible for creating/owning a new WebContents (a.k.a. pre-created
WebContents) to prerender a page for a new tab. The handles asks another
PrerenderHostRegistry associated with the pre-created WebContents to
start prerendering. The pre-created WebContents will have a blank
primary main frame (i.e., sitting at an initial empty document), and the
prerendered page will be in a separate RenderFrameHost in kPrerendering
state. The handle is retained by PrerenderHostRegistry on the triggering
WebContents (not the pre-created WebContents) so that the registry can
handle cancellation requests, for example, initiated by removal of the
speculation rules in the trigger page.

A notable point in this part is that the pre-created WebContents has a
tentative WebContentsDelegate during prerendering. This will be swapped
with a proper one during activation. This CL implements only the basic
functions of the delegate as a first step.

2) Activation
When a request to open a new tab/window comes from the trigger page,
the browser checks if there is a pre-created WebContents for navigation.
If it exists, the browser reuses the pre-created WebContents to handle
the request instead of creating a new WebContents. Then, the browser
starts the navigation sequence in the pre-created WebContents and run
prerender activation as usual.

> TODOs for follow-up CLs

Following things will be addressed by follow-up CLs:

- Support Precog for the prerender-in-new-tab mode.
- Support Extensions for the prerender-in-new-tab mode.
- Flesh out PrerenderNewTabHandle::WebContentsDelegateImpl and add test
  cases.

Change-Id: I3f3c7111b1da7a08ef4493ef4eeb23479c257d87
Bug: 1350676
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3890670
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075500}
NOKEYCHECK=True
GitOrigin-RevId: 3e25270476bf0b0c7e69684d4b53fb9f9b30e25c