When scrollbar changes avoid marking the ancestor chain in pre-paint
In CL:3613337 we were removing some code in
PaintLayer::UpdateLayerPositionRecursive() as it looks
it was not needed anymore.
However clusterfuzz has found a regression due to that change.

In order to fix the regression we're avoiding to mark the ancestor chain
in PaintLayerScrollableArea::ScrollbarVisibilityChanged()
when we're in pre-paint phase.

Bug: 1377634
Change-Id: Ie53f226bbc484cf6d49c454ab35ff246eb606891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4013540
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Manuel Rego <rego@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1075639}
NOKEYCHECK=True
GitOrigin-RevId: 50fa6378769cbb9f03145d7a43012c53eec49693