[devtools] Import web tests from devtools-frontend repo
When we moved DevTools to it's own repository we experimented with also
moving the DevTools related web tests. This turned out to be
infeasible as we would loose a great amount of coverage.

This CL imports the tests that are found only in devtools-frontend,
after which we can remove the web tests in DevTools.

Bug: 1393314
Change-Id: I9c6470ecb580fb3c67ef8d2e94a81513c0b92c1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055875
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075702}
NOKEYCHECK=True
GitOrigin-RevId: 45df4c5c49f6ecb0978c8e2eb13aa17fffac5e2d