Add enterprise policy to disable cross-origin non-visible iframe throttling.
This CL implements an enterprise policy to allow disabling of the
ThrottleDisplayNoneAndVisibilityHiddenCrossOriginIframes feature in
Blink.

Since the override requires providing a helper function directly in
third_party/blink/common, it seems simpler to not use
RuntimeEnabledFeatures for this feature, and instead just directly
declare it in blink::features. This requires small changes in the
tests, but overall makes the code easier to understand.

Bug: 1330592
Change-Id: I75eda216805864a266716d16d06fdb9d3d581cbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3966873
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1075923}
NOKEYCHECK=True
GitOrigin-RevId: b0bba18d143a3991930dca97e41c69883a729e00