PiP 2.0: Opening a PiP window from a PiP window is not allowed
This CL makes sure documentPictureInPicture.requestWindow() rejects
from a PiP window. Note that it wasn't possible already to open a
PiP window from a PiP window as a similar check already existed in the
browser side. The problem was that the promise returned by
documentPictureInPicture.requestWindow() was not rejected when it was
the case.

Spec: https://github.com/WICG/document-picture-in-picture/issues/13
Test: https://doc-pip.glitch.me
Bug: 1386162
Change-Id: I108b3b6edba8c595bfe0855141c034d5f9688dd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4056896
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Fr <beaufort.francois@gmail.com>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076109}
NOKEYCHECK=True
GitOrigin-RevId: 42da0956411540582ac2e7b6e628ac44aa5cb126