Only cache known-exposed properties in CSSStyleDeclaration
CSSStyleDeclaration has a cache which maps e.g. backgroundPositionY
to its associated CSSPropertyID. This cache is static, and is never
invalidated.

The problem is that the cache is populated with information based on
the ExecutionContext (and current runtime flags) used at the time we
happen to access backgroundPositionY for the first time. This means
that an ExecutionContext which enables a certain property via an
origin trial can "leak" that enabling for the whole process.

This CL fixes that by only adding known-exposed properties into the
cache.

Change-Id: I40ed950b7b0201df9c3fd8d9dfb22ae0be31a759
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055560
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076127}
NOKEYCHECK=True
GitOrigin-RevId: 1eb625cf62522f6a1e19afbf1629809d2c14d307