Add UseCounter for ServiceWorkerBypassFetchHandlerForMainResource
This CL adds a UseCounter to track the usage of
ServiceWorkerBypassFetchHandlerForMainResource feature.

This is for the future OriginTrials, and we're still not sure if we
should use the same metric for the upcoming Finch trials or not, but
this change is needed for the OriginTrials anyway.

Bug: 1371756
Change-Id: I837fc5ce05b986ed99f9308df89daefe12ca49de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4060121
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076760}
NOKEYCHECK=True
GitOrigin-RevId: 959864a67e535e82d35371cf873bda48ea983641