Improve HasIntrinsicSize() for StyleFetchedImage{,Set}
For SVG images, call GetIntrinsicSizingInfo() and then check the fields
of the returned structure. StyleImage should probably provide a way to
return the structure, but this will be a later step.

This allows BackgroundImageGeometry (et al) to compute a more correct
size in some cases.

Move SVGImage pAR 'none' sizing quirk to
SVGImage::GetIntrinsicSizingInfo. This way we preserve the quirk even
when not using SVGImage::ConcreteObjectSize.

In the tests:

 svg/wicd/test-scalable-background-image1.xhtml
 svg/zoom/page/zoom-svg-as-background-with-relative-size.html

the referenced image does not have intrinsic dimensions, so the snapped
positioning area will be used when resolving the image size.

Fixed: 935057
Change-Id: I03680ea2f2694500f09901bba613db275d0e2841
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4037347
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1076899}
NOKEYCHECK=True
GitOrigin-RevId: 314f2a7f28782ee867abb80a5cc08c617a5ffe0d