Tentative fix for CanvasAsyncBlobCreator clusterfuzz issue
When creating a blob, CanvasAsyncBlobCreator needs to use a non-gpu
image in order to produce the blob. Once the callback is called,
the blob creator will make an internal and temporary copy of the
(potentially) gpu backed image, and produce a cpu image to process it.
This could create some problems as, if the gpu-image stops existing
during the async the image can be deleted and we would be having a
null image to process.

This CL attempts to fix that by creating the unaccelerated copied
when initializing the CanvasAsyncBlobCreator, so it will have the
copy internally (in case that it was gpu backed) in case that there
is any issue with the gpu in order to process it.

This CL also adds a DCHECK in CanvasAsyncBlobCreator to catch this
issue with a DCHECK without having to rely on ASAN builds.

Bug: 1381881
Change-Id: Iac9b48e439885be3084d8c8df80f72bf70dcddcd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4009789
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Reviewed-by: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077016}
NOKEYCHECK=True
GitOrigin-RevId: ca864ccd8a2d227e88a40717f14f47c90631a703