WebNN: Create XNNPACK Subgraph and Runtime for MLGraphXnnpack
This CL introduces MLGraphXnnpack::CreateXnnSubgraphAndRuntime() method
that creates an XNNPACK Subgraph object as the representation of a WebNN
graph and creates an XNNPACK Runtime object from the Subgraph object for
the accelerated executions. It is called by
MLGraphXnnpack::BuildOnBackgroundThread() and runs in a worker thread
for asynchronous graph building. It can also run in the caller's thread
for synchronous graph build.

This CL initially creates an empty XNNPACK Subgraph object and its
Runtime object for testing purposes. The following CLs will continue to
implement this method that defines the Subgraph Values and Nodes for the
WebNN graph operands and operators in topological order.

Bug: 1273291
Change-Id: Ia5a8afebbe564152b475f39a773a657467f6e3d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3974030
Commit-Queue: ningxin hu <ningxin.hu@intel.com>
Reviewed-by: Jiewei Qian <qjw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077169}
NOKEYCHECK=True
GitOrigin-RevId: 602c7a5c3658b41f22cd8bbaa7a804acdb4fb0cb