Lower the priority of parkable string compression
Before, the task for `ParkableStringImpl::CompressInBackground` was
spawned with the default and highest `USER_BLOCKING` priority.
This is problematic as it competes with, e.g., WebAssembly compilation
in V8, which executes with the same priority in the same thread pool.
As a memory optimization, string compression is fine to spawn with the
lower `BEST_EFFORT` priority as well.

This improves scheduling for heavy WebAssembly applications running on
low-end devices, e.g., CPUs with just 2 cores without SMT.
Specifically for Photoshop on a low-end Chromebook, WebAssembly
compilation is now no longer interrupted by string compression.

Change-Id: I18c5d39767346c9599b52d262a8b6d4d56bb59d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061353
Reviewed-by: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077453}
NOKEYCHECK=True
GitOrigin-RevId: 6e92d75e5f1070f9d3fe6484f6ac2a8eba4a0aae