VTOnNav: Fix UAF that happens if the navigation request is deleted
We wait for the capture to ack before setting capture parameters.
However, during this, the navigation request can be deleted.

This patch fixes that by using a weak ptr on the navigation request.

Also fixes a couple of other DCHECKs that happened as a result of
testing this.

R=khushalsagar@chromium.org, alexmos@chromium.org

Bug: 1393564
Change-Id: I71e225eaa03fe8e74f9c07cc6deeee0da9a46c8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4061374
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Vladimir Levin <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077590}
NOKEYCHECK=True
GitOrigin-RevId: 4200b3628f3bf45133fefe67d0947f0f6899eb29