COM marshaling design change, isolate user/system interface registration
This CL introduces marshaling and registration logic to make different
(but binary-identical) COM interfaces for user and system respectively
to work seamlessly.

This CL also makes the changes for fully side-by-side interfaces for the
`Internal` updater interfaces in `updater_internal_idl.template`. The
non-internal updater interfaces in `updater_idl.template` will be added
in a follow-up CL.

The distinct interface IDs are suffixed with `User` and `System`
respectively. Most of the codebase uses the interface without the `User`
or `System` suffix, except the code that does the marshaling.

The interfaces with `User` and `System` suffixes have (and always need
to have) the exact same vtable and parameter layout as the interfaces
without the suffixes, which is what makes it possible for the marshaling
framework to think that it is marshaling the distinct `User` or `System`
interfaces, while the majority of the codebase thinks it is working with
the non-suffixed interfaces.

The `User` interfaces are only registered for the user case in HKCU,
while the corresponding `System` interfaces are only registered for the
system case in HKLM.

This isolates the installs from causing side-effects, and makes user and
system installs fully SxS.

Without this change, marshaling can load the typelib from `HKCU` for a
system install with UAC off, because when COM looks up registration,
entries in `HKCU` take priority over `HKLM` entries when UAC is off.

Bug: 1341471
Change-Id: I61a13c525eee4d3f22f92b82d8aed40eb5b74c74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4068183
Auto-Submit: S Ganesh <ganesh@chromium.org>
Commit-Queue: Xiaoling Bao <xiaolingbao@chromium.org>
Reviewed-by: Xiaoling Bao <xiaolingbao@chromium.org>
Commit-Queue: S Ganesh <ganesh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077639}
NOKEYCHECK=True
GitOrigin-RevId: a07f69bcecb82b749e156f45741c93d93744b947