Refactor WebmMuxer into a Muxer interface and implementation.
This change introduces the Muxer interface and refactors the WebmMuxer
to be an implementation of this interface. This creates room for future
implementations that support new containers, such as an Mp4Muxer.
More details about future plans can be found in this design doc:
https://microsoft-my.sharepoint-df.com/:w:/p/auorion/ERia8sV7FghKrCGzQ8NbENUBpYDq8tdVt3WFyOPhSaRWHQ

We also identify a couple places in the WebmMuxer where the
MediaRecorder spec (the primary consumer of the WebmMuxer) seems to
have been misinterpreted. We should discuss if we should fix these, or
have new implementations conform to the existing behavior.

Bug: 1072056
Change-Id: Ib26386ac28f6a8b28a782f41a40cea9d35dc3526
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3984226
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Markus Handell <handellm@google.com>
Commit-Queue: Austin Orion <auorion@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1077804}
NOKEYCHECK=True
GitOrigin-RevId: 0454d858a77e0d91f87c1cebae4a2095e8b7fb49