Deflake network requests and interactions trace event tests
The deflake consists in looking for the trace events specifically
related with the request for the HTML document and waiting for the DOM
to be loaded before stopping the trace. This way we make sure that all
trace events belong to the same network request and that every step in
the network request lifecycle is included.

As a drive by fix, the interactions trace events test is deflaked by using Page::loadEventFired to determine when the page is interactive.

Bug: 1383130
Change-Id: Ie6eb21a954773eb624527f329fcd39ea618d751f
Fixed: 1394220, 1394138
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4064120
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1077923}
NOKEYCHECK=True
GitOrigin-RevId: 4dd9f234c5ac7e11fd550518637a3c219ef2003e