user-notes: Ensure clean layout in ScrollIntoView
ScrollIntoView is called from the browser in response to a user click; it can happen at any time in the document lifecycle. Layout needs to
be clean in order to know where the target is on the page so this CL
ensures layout is up to date, preventing a DCHECK in debug builds.

Change-Id: I726136c417b4290ec4cb9ac620b52bc15bfb529d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4020921
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078094}
NOKEYCHECK=True
GitOrigin-RevId: a3f29e00bb59324e90fc4d4fac889b8d65597971