Shared Storage: Temp disable Resource Panel tests, update expectations.
We temporarily disable the DevTools Resource Panel web tests so that
the Shared Storage DevTools storage UX can land. We also update the
test expectations to include Shared Storage in the Resources Panel
so that they will be correct when we come back to re-enable the tests.

Bug: 1372007,1218540
Change-Id: I2ff95380ec824e5ef77cd77c0a2d5daf76dba86d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4071685
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078235}
NOKEYCHECK=True
GitOrigin-RevId: aa509f27295067e4b072324af20c94b11e0dc8ef