Update StorageKey::CreateForTesting to set ACB based on inputs
The StorageKey::CreateForTesting methods are useful for creating
third-party storage keys, but currently always set the ancestor
chain bit to kSameSite. We should set these to kCrossSite if the
origin and top-level site are cross-site since this is how it
would be for "real" storage keys.

Bug: None
Change-Id: If8fa472943cc8d58e82aa5d4ee7a66d92e854b7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4055220
Reviewed-by: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: Steven Bingler <bingler@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078240}
NOKEYCHECK=True
GitOrigin-RevId: 8c14680d21ccdce2753e3897eb9b7215972fb61a