[popover][anchor-position] Implement implicit anchoring (2/2)
- https://drafts.csswg.org/css-anchor-1/#implicit-anchor-element
- https://github.com/josepharhar/html/pull/2

This is the 2/2 patch implementing implicit anchoring for popovers,
focusing on layout-level changes.

With the info of which elements are used as implicit anchors, this
patch adds the implicit anchors into NG*AnchorQuery with LayoutObject
as keys, so that implicit anchor queries can be evaluated without an
anchor name.

At a more detailed level, now all the NG*AnchorQuery classes need to
maintain two hashmaps: one keyed by names for the named anchors, and
the other keyed by LayoutObjects for the implicit anchors. To reduce
code verbosity, the common structure of these classes is abstracted
into a base class NGAnchorQueryBase.

Bug: 1307772, 1380112
Change-Id: I9bdec5e3c08f9b9b20442503dedf32995643f037
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053342
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078270}
NOKEYCHECK=True
GitOrigin-RevId: 9328dcc2f1a0bba1fa5466cc30605b11b19076cc