Propagate the gesture for RemoteFrame::Navigate()
It’s unclear why we didn’t set gesture in some ResourceRequest
and/or propagate the gesture bit via RemoteFrame::Navigate().
(i.e. if this is to avoid exposing it cross-site, we can throttle
at a later point in the browser process / we already did this in
[1]). This CL sets this bit in the corresponding ResourceRequest
and thus fixes the behavior of:
1) download-in-ad-without-gesture intervention: we intend to
intervene downloads initiated and/or instantiated in ad iframe
without gesture, but the implementation was too strict because we
effectively treated cross-site navigation as no-gesture as the
gesture wasn’t propagated. (Things have been working fine as the
impact is too small / no one complained about this).
2) Sec-Fetch-User header: the intended behavior should be that
this header is sent for requests initiated by user activation, but
the header wasn’t sent as the gesture wasn’t set correctly.

In the future, we may also want to record metrics (e.g. whether
the landing page was initiated from a navigation with user gesture),
which would also benefit from having this bit propagated correctly.

[1] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/navigation_controller_impl.cc?q=%22false%20%2F*%20has_user_gesture%20*%2F%22

Bug: 1392168
Change-Id: Ifb2f9dbff544c1ffa32699ea0b96d0f41d930eb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4011119
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1078652}
NOKEYCHECK=True
GitOrigin-RevId: 5131e3df68fb579e624763da7b0eace288d0f4d9